Add use_times=False for simulations testing error
The modifications in 48bb35be753a23f20d77e5660a6742b3bbf5e31f revealed a current bug in the test suite: we add errors to a TS by adding extra mutations, but then expect to be able to get the time for a single site where there are mutliple mutations. Previously this bugged out to time = `None`, which was interpreted as "use frequency for time", hence times for sites with multiple mutations were allocated on freq, whereas other sites were using the msprime timescale. In all of these tests we should set use_times=False so all times are frequency-based