Back out of impute_missing option.
Including the impute_missing option seems like more trouble that it's
worth, as it adds a lot of complexity for what seems like a niche
requirement. It doesn't help us to test the imputation code in any way
which seems like the most important feature. Some handy utilities and
tests for splitting edges are removed here, which might be worth
disinterring in the future.